add Shiki fallback to LiveCodeBlock via Astro wrapper
