+++ src/content/posts/2025-09-22.paper-design-shaders.mdx
@@ -63,9 +63,9 @@ This one creates a flowing, organic background effect that works great for hero 
 
 Here's what that looks like in action:
 
-import LiveCodeBlock from '~/components/LiveCodeBlock';
+import LiveCodeBlock from '~/components/LiveCodeBlock.astro';
 
-<LiveCodeBlock client:load code={`import { MeshGradient } from '@paper-design/shaders-react' // ^0.0.55
+<LiveCodeBlock code={`import { MeshGradient } from '@paper-design/shaders-react' // ^0.0.55
 
 export default function App() {
   return (
